+++ mod.ts
@@ -11,4 +11,7 @@ export {
   type UseFullSoakOptions,
 } from "./src/useFullSoak.ts";
 export { getOrigin } from "./src/metastore.ts";
-export { setupDefaultFullsoakLogger } from "./src/utils.ts";
+export {
+  getFrameworkVersion,
+  setupDefaultFullsoakLogger,
+} from "./src/utils.ts";